requirements.txt
fastapi
uvicorn[standard]
pydantic
playwright
beautifulsoup4
lxml
orjson
dbt-duckdb
pytest
httpx
azure-storage-blob
duckdb
black
ruff
mypy
